move domain use cases to :domain
