+++ rss-generator/domain/src/main/kotlin/com/livefast/eattrash/rssgenerator/domain/FilePrinter.kt
@@ -1,0 +1,13 @@
+package com.livefast.eattrash.rssgenerator.domain
+
+/**
+ * Use case responsible for writing generated content to a persistent file.
+ */
+interface FilePrinter {
+    /**
+     * Writes the provided string content to file.
+     *
+     * @param content text content to be written.
+     */
+    fun execute(content: String)
+}
