+++ rss-generator/domain/src/main/kotlin/com/livefast/eattrash/rssgenerator/domain/FilePrinterImpl.kt
@@ -2,20 +2,7 @@ package com.livefast.eattrash.rssgenerator.domain
 
 import java.io.File
 
-/**
- * Use case responsible for writing generated content to a persistent file.
- */
-interface FilePrinter {
-    /**
-     * Writes the provided string content to file.
-     *
-     * @param content text content to be written.
-     */
-    fun execute(content: String)
-}
-
-class FilePrinterImpl : FilePrinter {
-
+internal class FilePrinterImpl : FilePrinter {
     override fun execute(content: String) {
         val destination = File(OUTPUT_FILE_PATH)
         destination.parentFile?.mkdirs()
