+++ rss-generator/domain/src/main/kotlin/com/livefast/eattrash/rssgenerator/domain/di/DomainModule.kt
@@ -9,11 +9,12 @@ import org.koin.dsl.module
 /**
  * DI module for the domain layer.
  */
-val domainModule = module {
-    single<RssGenerator> {
-        RssGeneratorImpl(dateUtils = get())
+val domainModule =
+    module {
+        single<RssGenerator> {
+            RssGeneratorImpl(dateUtils = get())
+        }
+        single<FilePrinter> {
+            FilePrinterImpl()
+        }
     }
-    single<FilePrinter> {
-        FilePrinterImpl()
-    }
-}
